+++ FE/src/components/Root.tsx
@@ -9,6 +9,7 @@ import NewIssue from "./NewIssue/NewIssue";
 import { useRecoilValue } from "recoil";
 import { LoginState } from "../atoms/atoms";
 import NotFound from "./Error/NotFound";
+import AuthRoute from "./Login/AuthRoute";
 
 const Root = () => {
   return (
@@ -26,7 +27,10 @@ const RouterComponent = () => {
       {/* 라우트 추가 부분 */}
       <Switch>
         <Route path="/" component={Login} exact />
-        <Route path="/issuelist" component={IssueList} />
+        <AuthRoute
+          path="/issuelist"
+          render={(props) => <IssueList {...props} />}
+        />
         <Route path="/newissue" component={NewIssue} />
         <Route path="/callback" component={Callback} />
         <Route component={NotFound} />
